+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add pattern types and modules
+    - `BoxPattern`
+    - `PurposePattern`
+    - `VersionPattern`
+    - `ZonePattern`
+    - `BucketPattern`
 
 ## 3.3.0 - 2020-01-14
 - Add function to access inner values